pyproject.toml
[project]
name = "datathon-fraude"
version = "0.1.0"
description = "ML Copilot para prevenção a fraude — Datathon Fase 05"
requires-python = ">=3.11"

dependencies = [
    # dados
    "sdv>=1.9.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "pandera>=0.18.0",
    "dvc>=3.0.0",

    # ml
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "torch>=2.2.0",
    "shap>=0.44.0",

    # mlflow
    "mlflow>=2.10.0",

    # llm e agente
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-groq>=0.1.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.5.0",

    # api
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",

    # observabilidade
    "evidently>=0.4.0",
    "langfuse>=2.0.0",

    # segurança
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 150
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
